[account config] Remove unused functions and variables
