+++ types/src/account_config/constants/account.rs
@@ -2,10 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::account_config::constants::CORE_CODE_ADDRESS;
-use move_core_types::{
-    identifier::{IdentStr, Identifier},
-    language_storage::ModuleId,
-};
+use move_core_types::{identifier::Identifier, language_storage::ModuleId};
 use once_cell::sync::Lazy;
 
 pub const ACCOUNT_MODULE_NAME: &str = "LibraAccount";
@@ -17,17 +14,3 @@ pub static ACCOUNT_MODULE_IDENTIFIER: Lazy<Identifier> =
 /// The ModuleId for the Account module.
 pub static ACCOUNT_MODULE: Lazy<ModuleId> =
     Lazy::new(|| ModuleId::new(CORE_CODE_ADDRESS, ACCOUNT_MODULE_IDENTIFIER.clone()));
-
-// Payment Events
-static SENT_EVENT_NAME: Lazy<Identifier> =
-    Lazy::new(|| Identifier::new("SentPaymentEvent").unwrap());
-static RECEIVED_EVENT_NAME: Lazy<Identifier> =
-    Lazy::new(|| Identifier::new("ReceivedPaymentEvent").unwrap());
-
-pub fn sent_event_name() -> &'static IdentStr {
-    &*SENT_EVENT_NAME
-}
-
-pub fn received_event_name() -> &'static IdentStr {
-    &*RECEIVED_EVENT_NAME
-}
