+++ types/src/account_config/constants/mod.rs
@@ -2,21 +2,13 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod account;
-pub mod account_limits;
-pub mod account_type;
 pub mod addresses;
 pub mod coins;
-pub mod debug;
 pub mod event;
-pub mod lbr;
 pub mod libra;
 
 pub use account::*;
-pub use account_limits::*;
-pub use account_type::*;
 pub use addresses::*;
 pub use coins::*;
-pub use debug::*;
 pub use event::*;
-pub use lbr::*;
 pub use libra::*;
